- tool seo : can disable opengraph

+++ core/tools/seo/custom-fields/seo.php
@@ -4,7 +4,7 @@
  * @author Sébastien Chandonay www.seb-c.com / Cyril Tissot www.cyriltissot.com
  * License: GPL2
  * Text Domain: woodkit
- * 
+ *
  * Copyright 2016 Sébastien Chandonay (email : please contact me from my website)
  *
  * This program is free software; you can redistribute it and/or modify
@@ -28,6 +28,7 @@ define('SEO_OPTIONS_NAME', 'seo-options');
 define('SEO_CUSTOMFIELD_METATITLE', 'seo-meta-title');
 define('SEO_CUSTOMFIELD_METADESCRIPTION', 'seo-meta-description');
 define('SEO_CUSTOMFIELD_METAKEYWORDS', 'seo-meta-keywords');
+
 define('SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE', 'seo-meta-og-title');
 define('SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION', 'seo-meta-og-description');
 define('SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE', 'seo-meta-og-image');
@@ -62,40 +63,44 @@ function seo_save_post($post_id){
 		}else{
 			delete_post_meta($post_id, SEO_CUSTOMFIELD_METATITLE);
 		}
-		
+
 		// SEO_CUSTOMFIELD_METADESCRIPTION
 		if (!empty($_POST[SEO_CUSTOMFIELD_METADESCRIPTION])){
 			update_post_meta($post_id, SEO_CUSTOMFIELD_METADESCRIPTION, sanitize_text_field($_POST[SEO_CUSTOMFIELD_METADESCRIPTION]));
 		}else{
 			delete_post_meta($post_id, SEO_CUSTOMFIELD_METADESCRIPTION);
 		}
-		
+
 		// SEO_CUSTOMFIELD_METAKEYWORDS
 		if (!empty($_POST[SEO_CUSTOMFIELD_METAKEYWORDS])){
 			update_post_meta($post_id, SEO_CUSTOMFIELD_METAKEYWORDS, sanitize_text_field($_POST[SEO_CUSTOMFIELD_METAKEYWORDS]));
 		}else{
 			delete_post_meta($post_id, SEO_CUSTOMFIELD_METAKEYWORDS);
 		}
-		
-		// SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE
-		if (!empty($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE])){
-			update_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE, sanitize_text_field($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE]));
-		}else{
-			delete_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE);
-		}
-		
-		// SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION
-		if (!empty($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION])){
-			update_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION, sanitize_text_field($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION]));
-		}else{
-			delete_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION);
-		}
-		
-		// SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE
-		if (!empty($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE])){
-			update_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE, sanitize_text_field($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE]));
-		}else{
-			delete_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE);
+
+		$opengraph_active = woodkit_get_option("tool-seo-opengraph-active");
+		if (!empty($opengraph_active) && $opengraph_active == "on"){
+
+			// SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE
+			if (!empty($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE])){
+				update_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE, sanitize_text_field($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE]));
+			}else{
+				delete_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_TITLE);
+			}
+
+			// SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION
+			if (!empty($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION])){
+				update_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION, sanitize_text_field($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION]));
+			}else{
+				delete_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_DESCRIPTION);
+			}
+
+			// SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE
+			if (!empty($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE])){
+				update_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE, sanitize_text_field($_POST[SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE]));
+			}else{
+				delete_post_meta($post_id, SEO_CUSTOMFIELD_META_OPENGRAPH_IMAGE);
+			}
 		}
 	}
 }
@@ -161,7 +166,7 @@ endif;
 if (!function_exists("seo_registered_taxonomy")):
 /**
  * add seo on woodkit taxonomy extra fields
- */
+*/
 function seo_registered_taxonomy($taxonomy){
 	$seo_exludes_taxonomies = array("post_tag", "nav_menu", "link_category", "post_format");
 	if (!in_array($taxonomy, $seo_exludes_taxonomies)){
